2nd round of mingw fixes
git-svn-id: https://svn.kvirc.de/svn/trunk/kvirc@2978 17fca916-40b9-46aa-a4ea-0a15b648b75c

+++ src/modules/reguser/dialog.cpp
@@ -200,7 +200,7 @@ QSize KviRegisteredUsersDialogItemDelegate::sizeHint( const QStyleOptionViewItem
 		//users
 // 		KviRegisteredUsersDialogItem *it=(KviRegisteredUsersDialogItem*)item;
 
-		return QSize(300, (int) LVI_ICON_SIZE+2.0*LVI_BORDER);
+		return QSize(300, LVI_ICON_SIZE+2*LVI_BORDER);
 	}
 }
